+++ mrmustard/representations/data/MatVecData.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from mrmustard.representations import MatVecData
+from mrmustard.representations.data import Data
 
-class SimplecticData(MatVecData):
+class MatVecData(Data):
 
-    def __init__(self) -> SimplecticData:
+    def __init__(self) -> MatVecData:
         super().__init__()
