+++ mrmustard/representations/data/QPolyData.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from mrmustard.representations import MatVecData
+from mrmustard.representations.data import MatVecData
 
 class QPolyData(MatVecData):
